+++ packages/client/src/modules/fields/component/PopoverNester/subformPopoverComponentFieldSettings.tsx
@@ -27,7 +27,9 @@ const allowMultiple: any = {
     return {
       title: t('Allow multiple'),
       checked:
-        fieldSchema['x-component-props']?.multiple === undefined ? true : fieldSchema['x-component-props'].multiple,
+        (fieldSchema['x-component-props'] || {}).multiple === undefined
+          ? true
+          : fieldSchema['x-component-props'].multiple,
       onChange(value) {
         const schema = {
           ['x-uid']: fieldSchema['x-uid'],